refactor(issue): integrate workflow runs into IssueWorkflowRuns component and remove client component
- Migrate workflow run fetching to the IssueWorkflowRuns component using SWR for real-time updates.
- Pass initial workflow runs as props to the component.
- Remove the deprecated IssueWorkflowRunsClient component to streamline the codebase.
